Move AuthenticatedUsersPageContainer
