+++ packages/app-root/src/app/users/[login]/groups/MyGroupsContainer.js
@@ -4,7 +4,7 @@ import { MyGroups } from '@zooniverse/user'
 import { useContext } from 'react'
 
 import { PanoptesAuthContext } from '../../../../contexts'
-import AuthenticatedUsersPageContainer from '../components/AuthenticatedUsersPageContainer'
+import AuthenticatedUsersPageContainer from '../../../../components/AuthenticatedUsersPageContainer'
 
 function MyGroupsContainer({
   login
