+++ packages/app-root/src/app/users/[login]/stats/UserStatsContainer.js
@@ -4,7 +4,7 @@ import { UserStats } from '@zooniverse/user'
 import { useContext } from 'react'
 
 import { PanoptesAuthContext } from '../../../../contexts'
-import AuthenticatedUsersPageContainer from '../components/AuthenticatedUsersPageContainer'
+import AuthenticatedUsersPageContainer from '../../../../components/AuthenticatedUsersPageContainer'
 
 function UserStatsContainer({
   login
